+++ src/pages/administration/teachers/index.vue
@@ -271,11 +271,6 @@ export default {
 		...mapState("search", {
 			searchResult: "searchResult",
 		}),
-		permissionFilteredTableActions() {
-			return this.tableActions.filter((action) =>
-				action.permission ? this.$_userHasPermission(action.permission) : true
-			);
-		},
 		tableData: {
 			get() {
 				if (this.takeOverTableData) return this.searchData;
@@ -283,30 +278,54 @@ export default {
 			},
 		},
 		filteredActions() {
+			let editedActions = this.tableActions;
+
+			// filter actions by permissions
+			editedActions = this.tableActions.filter((action) =>
+				action.permission ? this.$_userHasPermission(action.permission) : true
+			);
+
 			// filters out the QR bulk action is user is not an admin
 			if (!this.user.roles.some((role) => role.name === "administrator")) {
-				return this.permissionFilteredTableActions.filter(
+				editedActions = editedActions.filter(
 					(action) =>
 						action.label !==
 						this.$t("pages.administration.teachers.index.tableActions.qr")
 				);
 			}
 
-			return this.permissionFilteredTableActions;
+			// filter the delete action if school is external
+			if (this.school.isExternal) {
+				editedActions = editedActions.filter(
+					(action) =>
+						action.label !==
+						this.$t("pages.administration.teachers.index.tableActions.delete")
+				);
+			}
+
+			return editedActions;
 		},
 		filteredColumns() {
-			// filters out edit/consent column if school is external or if user is not an admin
+			let editedColumns = this.tableColumns;
+			// filters out edit column if school is external or if user is not an admin
 			if (
 				this.school.isExternal ||
 				!this.user.roles.some((role) => role.name === "administrator")
 			) {
 				return this.tableColumns.filter(
 					// _id field sets the edit column
-					(col) => col.field !== "_id" && col.field !== "consentStatus"
+					(col) => col.field !== "_id"
 				);
 			}
 
-			return this.tableColumns;
+			// filters out the consent column if ADMIN_TABLES_DISPLAY_CONSENT_COLUMN env is disabled
+			if (!process.env["ADMIN_TABLES_DISPLAY_CONSENT_COLUMN"]) {
+				editedColumns = editedColumns.filter(
+					(col) => col.field !== "consentStatus"
+				);
+			}
+
+			return editedColumns;
 		},
 		schoolInternallyManaged() {
 			return !this.school.isExternal;
@@ -494,6 +513,13 @@ export default {
 				});
 			}, 400);
 		},
+	},
+	head() {
+		return {
+			title: `${this.$t("pages.administration.teachers.index.title")} - ${
+				this.$theme.short_name
+			}`,
+		};
 	},
 };
 </script>