+++ CHANGELOG.md
@@ -10,11 +10,18 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ## 26.1.0
 
+### Added
+
+- SC-8916 - set page titles to some administration pages
+- SC-8681 - added ADMIN_TABLES_DISPLAY_CONSENT_COLUMN as an environment variable to display/hide the NAT consent column
+
 ### Fixed
 
 - SC-8518 - fixed layout issues of checkbox
 - SC-8881 - fixed double roles user (admin, teacher) not being able to edit other teachers
 - SC-8519 - Admin Table Bugfix - Set current date for filter, without input; Error handling parsing dates
+- SC-8878 - Fixing invalid date on filter chips, setting default filter start and end day to 1900-01-01 and 2099-12-31
+- SC-8910 - fixed delete action being available when school is external
 
 ### Removed